requirements.txt
python-pptx
notebook
python-dotenv
Jinja2
